requirements.txt
streamlit==1.39.0
langchain==0.3.3
langchain-core==0.3.10
langchain-google-genai==1.0.6
python-dotenv==1.0.1